+++ src/main/java/subway/domain/path/JgraphtPathFinder.java
@@ -1,39 +1,47 @@
-package subway.domain;
+package subway.domain.path;
 
 import java.util.List;
 import java.util.stream.Collectors;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
+import subway.domain.Distance;
+import subway.domain.Line;
+import subway.domain.Section;
+import subway.domain.Station;
+import subway.domain.Subway;
 import subway.exception.StationNotFoundException;
 
-public class PathFinder {
+public class JgraphtPathFinder implements PathFinder {
 
     private final Subway subway;
 
-    public PathFinder(final Subway subway) {
+    public JgraphtPathFinder(final Subway subway) {
         this.subway = subway;
     }
 
-    public Path findShortestPath(final String startStationName, final String endStationName) {
-        if (!isExistStation(startStationName) || !isExistStation(endStationName)) {
-            throw new StationNotFoundException();
-        }
+    @Override
+    public Path find(final String startStationName, final String endStationName) {
+        validateExist(startStationName, endStationName);
 
         final WeightedMultigraph<String, DefaultWeightedEdge> graph = makeWeightGraph();
         final DijkstraShortestPath<String, DefaultWeightedEdge> dijkstraShortestPath =
                 new DijkstraShortestPath<>(graph);
         final List<String> pathWithStationName = dijkstraShortestPath.getPath(startStationName, endStationName)
                 .getVertexList();
+        final List<Station> stations = nameToStation(pathWithStationName);
         final double distance = dijkstraShortestPath.getPathWeight(startStationName, endStationName);
-        final List<Station> stations = pathWithStationName.stream()
-                .map(Station::new)
-                .collect(Collectors.toList());
 
         return new Path(stations, new Distance((int) distance));
     }
 
-    private Boolean isExistStation(final String stationName){
+    private void validateExist(String startStationName, String endStationName) {
+        if (!isExistStation(startStationName) || !isExistStation(endStationName)) {
+            throw new StationNotFoundException();
+        }
+    }
+
+    private Boolean isExistStation(final String stationName) {
         return subway.getLines().stream()
                 .anyMatch(line -> line.hasStation(stationName));
     }
@@ -67,5 +75,11 @@ public class PathFinder {
         if (!graph.containsVertex(stationName)) {
             graph.addVertex(stationName);
         }
+    }
+
+    private List<Station> nameToStation(final List<String> pathWithStationName) {
+        return pathWithStationName.stream()
+                .map(Station::new)
+                .collect(Collectors.toList());
     }
 }
