+++ src/main/java/subway/domain/path/PathFinder.java
@@ -1,0 +1,6 @@
+package subway.domain.path;
+
+public interface PathFinder {
+
+    Path find(final String startStationName, final String endStationName);
+}
